requirements.txt
# No external dependencies. Stdlib-only by design.
fastapi
httpx
uvicorn[standard]
PyMuPDF
openpyxl
playwright
psycopg2-binary
sse-starlette
google-auth
google-auth-oauthlib
